fix: widen telegram_user_id to BigInteger
Telegram user IDs can exceed 32-bit INTEGER range (e.g. 8603580375).
The column was declared as Integer which caused psycopg NumericValueOutOfRange
on bootstrap_user() lookup in the migration container.

Change UserORM.telegram_user_id from Integer to BigInteger.
The schema-wipe migration (0014) recreates the table so no ALTER is needed.

+++ packages/storage/src/helm_storage/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from uuid import uuid4
 
-from sqlalchemy import JSON, Boolean, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
+from sqlalchemy import JSON, BigInteger, Boolean, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.sql import func
 
@@ -12,7 +12,7 @@ class UserORM(Base):
     __tablename__ = "users"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    telegram_user_id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False)
+    telegram_user_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False)
     display_name: Mapped[str | None] = mapped_column(String(255))
     timezone: Mapped[str] = mapped_column(String(64), nullable=False, default="UTC")
     created_at: Mapped[datetime] = mapped_column(
